requirements.txt
# Core dependencies
langchain>=0.3.0
langgraph>=0.2.0
deepagents>=0.1.0

# LLM providers
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Data processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# Persistence
langgraph-checkpoint-postgres>=2.0.0
psycopg2-binary>=2.9.0

# Optional: For PDF parsing
pypdf>=4.0.0